Module Autolog.Log_Integer_Backward

Parameters

module C : INTEGER_CONVERSION

module F : sig ... end

Signature

val itimes : Z.t -> (C.integer, C.integer) C.Arity.ar1

Multiply an integer by a constant

val iadd : (C.integer, C.integer, C.integer) C.Arity.ar2

val imul : (C.integer, C.integer, C.integer) C.Arity.ar2

val idiv : (C.integer, C.integer, C.integer) C.Arity.ar2

This is truncated (C99-like) integer division

val imod : (C.integer, C.integer, C.integer) C.Arity.ar2

val ishl : (C.integer, C.integer, C.integer) C.Arity.ar2

val ishr : (C.integer, C.integer, C.integer) C.Arity.ar2

val iand : (C.integer, C.integer, C.integer) C.Arity.ar2

Bitwise and, where negative integers are seen as prefixed by infinite ones

val ior : (C.integer, C.integer, C.integer) C.Arity.ar2

Bitwise or, where negative integers are seen as prefixed by infinite ones

val ixor : (C.integer, C.integer, C.integer) C.Arity.ar2

val isub : (C.integer, C.integer, C.integer) C.Arity.ar2

val ieq : (C.integer, C.integer, C.boolean) C.Arity.ar2

val ile : (C.integer, C.integer, C.boolean) C.Arity.ar2