Module Conversions.Convert_Block_Forward

Parameters

module C : Conversion

module F : sig ... end

Signature

type boolean = F.boolean

type offset = F.offset

type value = F.value

type block = F.block

val sizeof : (block, offset) C.To_Arity.ar1

Size of a block in bytes

val concat : (block, block, block) C.To_Arity.ar2

Concatenates two blocks

val load : size:Units.In_bits.t -> (block, offset, value) C.To_Arity.ar2

Loads (extracts) a value of a fixed size at a given index from a block

val store : size:Units.In_bits.t -> (block, offset, value, block) C.To_Arity.ar3

Stores (writes) a fixed size value of a given index in a block

val binary_to_block : size:Units.In_bits.t -> (value, block) C.To_Arity.ar1

Converts a fixed size value to a block