Parameter Pretty.M

type 'a t

The term to print.

type 'a pack

Packing of the arguments

val pretty : Format.formatter -> 'a t -> unit

Term pretty printer

val extract1 : 'a ar1 pack -> 'a t

Extract the argument when there is one argument in a term.

val extract2 : ('a, 'b) ar2 pack -> 'a t * 'b t

Extract the arguments when there are two arguments in a term.