Module Sig.Forward_Arity

Standard arities for forward transfer functions: given the arguments, return the results. These match the arities of the concrete functions they represent (but with concrete types substituted for their abstract counterparts).

type 'r ar0 = 'r

type ('a, 'r) ar1 = 'a -> 'r

type ('a, 'b, 'r) ar2 = 'a -> 'b -> 'r

type ('a, 'b, 'c, 'r) ar3 = 'a -> 'b -> 'c -> 'r