Module type Sig.BLOCK_BACKWARD

type boolean

type offset

type value

type block

module Arity : ARITY

val sizeof : (block, offset) Arity.ar1

Size of a block in bytes

val concat : (block, block, block) Arity.ar2

Concatenates two blocks

val load : size:Units.In_bits.t -> (block, offset, value) Arity.ar2

Loads (extracts) a value of a fixed size at a given index from a block

val store : size:Units.In_bits.t -> (block, offset, value, block) Arity.ar3

Stores (writes) a fixed size value of a given index in a block

val binary_to_block : size:Units.In_bits.t -> (value, block) Arity.ar1

Converts a fixed size value to a block