Module type Sig.INTEGER_FORWARD

include INTEGER_FORWARD_MIN

include INTEGER_BACKWARD

type integer

type boolean

module Arity : ARITY

val itimes : Z.t -> (integer, integer) Arity.ar1

Multiply an integer by a constant

val iadd : (integer, integer, integer) Arity.ar2

val imul : (integer, integer, integer) Arity.ar2

val idiv : (integer, integer, integer) Arity.ar2

This is truncated (C99-like) integer division

val imod : (integer, integer, integer) Arity.ar2

val ishl : (integer, integer, integer) Arity.ar2

val ishr : (integer, integer, integer) Arity.ar2

val iand : (integer, integer, integer) Arity.ar2

Bitwise and, where negative integers are seen as prefixed by infinite ones

val ior : (integer, integer, integer) Arity.ar2

Bitwise or, where negative integers are seen as prefixed by infinite ones

val ixor : (integer, integer, integer) Arity.ar2

val isub : (integer, integer, integer) Arity.ar2

val ieq : (integer, integer, boolean) Arity.ar2

val ile : (integer, integer, boolean) Arity.ar2

val iconst : Z.t -> integer Arity.ar0

Integer constant

val zero : integer Arity.ar0

These can be defined from the others, but it may be more efficient to implement them directly (no need to build temporary values...). They are also often convenient to use directly.

val one : integer Arity.ar0